backend/pyproject.toml
[project]
name = "ios-chat-agent-template-backend"
version = "0.1.0"
description = "FastAPI backend for the iOS chat agent template."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "openai-agents>=0.13.2,<0.14.0",
  "pydantic-settings>=2.11.0,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<0.29.0",
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
